requirements.txt
Django==1.7.1
Pillow==2.6.1
argparse==1.2.1
beautifulsoup4==4.3.2
django-app-namespace-template-loader==0.2
django-blog-zinnia==0.15
django-contrib-comments==1.5
django-filebrowser==3.5.7
django-grappelli==2.5.6
django-mptt==0.6.1
django-tagging==0.3.4
django-tinymce==1.5.3
django-xmlrpc==0.1.5
psycopg2==2.5.3
pyparsing==2.0.3
pytz==2014.10
six==1.8.0
sorl-thumbnail==12.2
wsgiref==0.1.2
zinnia-wysiwyg-tinymce==1.2
